% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggTS.R
\name{aggTS}
\alias{aggTS}
\title{Aggregate data.frame time series to other time periods}
\usage{
aggTS(timeseries_data, newPeriod, fun = "sum", time_offset = 0,
  conserve_columns = NA)
}
\arguments{
\item{timeseries_data}{data.frame, formatted with columns [datetime, values].}

\item{newPeriod}{charater string, options are "hourly", "daily", "weekly".}

\item{fun}{character string, indicating aggregation function.}

\item{time_offset}{integer, indicating the offset to include in the aggregation.
Units are with respect to declared new period.}
}
\value{
test
}
\description{
test
}
\details{
Convention for setting dates corresponding to new period: POSIXct is
always used from the LAST hour, day, etc. of the aggregation period.
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
