% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_facTopo.R
\name{calc_facTopo}
\alias{calc_facTopo}
\title{Calculate topographic gravity factor}
\usage{
calc_facTopo(site_name, grid_extent, grid3d_discr, grid3d_depth,
  topo_flat = F, pillar_correction = T, dir_input_DEM,
  DEM_fileType = "ascii")
}
\arguments{
\item{grid_extent}{vector, giving as first argument x and as second y grid extent in [m]}

\item{grid3d_discr}{data.frame(x = 100, y = 100, z = 0.5)}

\item{grid3d_depth}{c(-10, 0) # min, max}

\item{dir_input}{directory of DEM file}
}
\value{
Returns the topographic conversion factor which represents
the gravity effect resulting from 1 mm of vertical water storage change.
This factor is used to convert mm of water from precipitation or evapotranspiration
measurements, etc. to gravity units [nm/s²].
}
\description{
This is used to transform units: [mm] to [nm/s²]
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
