% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gravityGrid.R
\name{create_gravityGrid}
\alias{create_gravityGrid}
\title{test}
\usage{
create_gravityGrid(DEM_input_file, dir_input_DEM, SG_coordinates,
  grid_discretization, grid_depth, range_coords_x = NA, range_coords_y = NA,
  radius_inner = NA, radius_outer = NA, ...)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
