% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nestedGravityGrid.R
\name{create_nestedGravityGrid}
\alias{create_nestedGravityGrid}
\title{Create a gravity component grid out of 3 spatially different discretized grids}
\usage{
create_nestedGravityGrid(DEM_input_files, DEM_dir, SGloc, grid_discretizations,
  grid_vertical, threshold_radi, correct_SGpillar = NA)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
