% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_from_gleam.R
\name{ET_from_gleam}
\alias{ET_from_gleam}
\title{Actual Evapotranspiration data (daily) from GLEAM}
\usage{
ET_from_gleam(input_dir, filenames, lon, lat, output_cumsum = F,
  plotting = T)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
Datasets are provided by: 
https://www.gleam.eu/
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
