% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ETavgs.R
\name{extract_ETavgs}
\alias{extract_ETavgs}
\title{Compare differences of value at midnight to daily diferences from actual time series}
\usage{
extract_ETavgs(data_observed, data_reference, topoFactor, avg_start, avg_end,
  plotting = FALSE, ts_start = NA, ts_end = NA)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
observed data input: hourly (cumsums), actual / reference input: daily (difs)
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
