% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facTopo_from_grid.R
\name{facTopo_from_grid}
\alias{facTopo_from_grid}
\title{Calculate topographic gravity factor}
\usage{
facTopo_from_grid(gravity_component_grid3d, ...)
}
\arguments{
\item{gravity_component_grid3d}{data.frame, supplying the gravity grid.
Needs columns x, y, z, Depth, gcomp (gravity component value).}
}
\value{
Returns the topographic conversion factor which represents
the gravity effect resulting from 1 mm of vertical water storage change.
This factor is used to convert mm of water from precipitation or evapotranspiration
measurements, etc. to gravity units [nm/s²].
}
\description{
This is used to transform units: [mm] to [nm/s²]
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
