% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComponentCombinations.R
\name{getComponentCombinations}
\alias{getComponentCombinations}
\title{Get the names of all gravity components, used in combinations}
\usage{
getComponentCombinations(site_name, input_dir, component_name = NA)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
Get a data.frame with all possible combinations of components for gravity corrections,
included in the corresponding folder.
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
