% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFluxNet_ET.R
\name{getFluxNet_ET}
\alias{getFluxNet_ET}
\title{Extract and format evapotranspiration data from FluxNet}
\usage{
getFluxNet_ET(site_name, input_dir, output_cumsum, plotting = F,
  show_raw = F, EB_correction = T, QF_correction = F, ...)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
