% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStationParam.R
\name{getStationParam}
\alias{getStationParam}
\title{Extract parameter corresponding to a station (location)}
\usage{
getStationParam(Site, Param, input_dir = "./")
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
