% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_ts.R
\name{pick_ts}
\alias{pick_ts}
\title{Construct a time series out of different combinations of components}
\usage{
pick_ts(input_dir, site_name, variation, timePeriod_start, timePeriod_end,
  includeET = FALSE, gravityUnits = TRUE, cummulative = TRUE, facTopo,
  ...)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
