% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_from_esrl.R
\name{precip_from_esrl}
\alias{precip_from_esrl}
\title{Precipitation data (daily) from ESRL (NOAA)}
\usage{
precip_from_esrl(input_dir, filenames, lon, lat, output_cumsum = F,
  plotting = T)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
Datasets are provided by: 
https://www.esrl.noaa.gov/psd/data/gridded/data.cpc.globalprecip.html
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
