% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip_from_GPM.R
\name{precip_from_GPM}
\alias{precip_from_GPM}
\title{Precipitation data (30 minutes) from GPM (NASA)}
\usage{
precip_from_GPM(input_dir, filenames, var_name = NA, lon, lat,
  output_cumsum = F, plotting = T)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
Datasets are provided by: 
https://www.esrl.noaa.gov/psd/data/gridded/data.cpc.globalprecip.html
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
