% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravtiyGrid_helper.R
\name{remove_outsideRadius}
\alias{remove_outsideRadius}
\title{removes all DEM points outside of a certain radius}
\usage{
remove_outsideRadius(grid_input, gloc, max_rad)
}
\arguments{
\item{grid_input}{grid from where to limit gcomp to a certain radius. Has to be data.frame with columns (x,y,z).}

\item{gloc}{location of gravity sensor (x,y,z).}

\item{max_rad}{radius of exclusion in [meters].}
}
\description{

}
\details{
missing
}
\examples{
example MISSING
}
\references{
Marvin Reich (2014), mreich@gfz-potsdam.de
}
