% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{RMSE of simulated and observed vectors}
\usage{
rmse(ts_sim, ts_obs)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
