% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif_to_ascii.R
\name{tif_to_ascii}
\alias{tif_to_ascii}
\title{tif_to_ascii: converts geoTiff fromat raster data into arc ascii format}
\usage{
tif_to_ascii(file_name_in, file_name_out, input_dir, output_dir)
}
\arguments{
\item{test}{test}
}
\value{
test
}
\description{
test
}
\details{
missing
}
\examples{
missing
}
\references{
Marvin Reich (2018), mreich@posteo.de
}
